Add a module for shared test utils
