+++ common-test-utils/README.md
@@ -1,0 +1,3 @@
+## :common-test-utils
+
+This internal module contains shared test utilities.
