+++ dataframe-csv/build.gradle.kts
@@ -35,6 +35,7 @@ dependencies {
     testImplementation(libs.kotestAssertions) {
         exclude("org.jetbrains.kotlin", "kotlin-stdlib-jdk8")
     }
+    testImplementation(projects.commonTestUtils)
 }
 
 benchmark {
