+++ dataframe-excel/build.gradle.kts
@@ -28,6 +28,7 @@ dependencies {
     testImplementation(libs.kotestAssertions) {
         exclude("org.jetbrains.kotlin", "kotlin-stdlib-jdk8")
     }
+    testImplementation(projects.commonTestUtils)
 }
 
 kotlinPublications {
