+++ dataframe-jdbc/build.gradle.kts
@@ -33,6 +33,7 @@ dependencies {
         exclude("org.jetbrains.kotlin", "kotlin-stdlib-jdk8")
     }
     testImplementation(libs.hikaricp)
+    testImplementation(projects.commonTestUtils)
 }
 
 kotlinPublications {
