+++ settings.gradle.kts
@@ -28,3 +28,4 @@ include("dataframe-geo-jupyter")
 include("dataframe-openapi-generator")
 include("core")
 include("dataframe-compiler-plugin-core")
+include("common-test-utils")
